Added ability to update diagram from last used view
As an extension to issue #93, this commit introduces
changes to allow updating the diagram based on the
view that PlantUML last used (i.e., the currentPart)
instead of the view that is currently in focus. To make
this work, PlantUML now ignores its own view (i.e., does
not set the view in which it renders the diagram as the
current part) and has a new method
updateDiagramTextFromCurrentPart() which updates the
diagram based on the currently set currentPart.